requirements.txt
boto==2.42.0
codegen==1.0
Django==1.10
django-appconf==1.0.2
django-disqus==0.5
django-imagekit==3.3
django-pipeline==1.6.8
django-storages==1.4.1
django-tagging==0.4.3
mock==2.0.0
patterns==0.3
pbr==1.10.0
pilkit==1.1.13
Pillow==3.3.0
six==1.10.0
